Moved some check & checkmate logic to Piece. Refactored.
I wanted us to be able to easily reference the piece that
caused check and the piece that caused checkmate.
Moving the logic for causing check and checkmate to
the Piece class meant I could also isolate methods for
enemy_king and friendly_king, since they are both pieces.

There is one failing test still. I believe it fails because
legal_move doesn't catch if the move leaves the friendly
king open to attack. But it could be for another reason.