Added Michelle's tests for checkmate, tried for green.
At this point, there are two failing tests, which expose some
buggy logic that we still have to sort out.

For instance, in the King class, I discovered that the under_attack?
method I wrote only applies to the current state of the board,
before the king (or other piece) has moved. We still need a way
of checking the potential under_attack? state after a piece has moved.
Especially in the cases where that move occured in a straight path
from the attacking piece...

Perhaps we could invoke the straight_obstruction_array between the
attacking piece and the square we want to move our piece to, and check
to see if our piece is currently in that path? If it is in that path,
then we could say that the destination square would be under_attack.