Continued to work on failing tests.
In king.rb, the can_move_out_of_check? method isn't working.
It returns true when it should return false.
(see the test for it 'should return true if king is in checkmate'
in game_spec.rb, line 161, for an example)

I added a ternary operator to line 15 in king.rb,
hoping to catch the case where the king considers moving away
from the attacking piece, but along the same path that the
attacking piece is attacking. In this case, the king would have
been blocking the destination square from attack, but he
would remain under attack if he moved there.

Something about my conditional logic is failing here.
And I'm annoyed at the ternary operator. I used it to make
my code more concise, but there are a lot of double
negatives that are confusing.