Flipped default game setup. Simplified check & checkmate methods.
I changed the board setup so that white starts from the bottom,
and black starts from the top. Most of the chess examples I found
online had this setup, and this should make it easier for us
to add future rspec tests, because we can base them on online
examples directly without having to invert the is_white values.

The setup matters mostly because pawns aren't allowed to move
backwards, so we need to define the backwards movement based on
where the pieces began.

I had previously defined check and checkmate based on whether an
attacking_piece (defined as the piece that moved last) was
causing check or checkmate. Unfortunately, defining the
attacking_piece as the piece which moved last is faulty logic.
Sometimes, you could move a piece to reveal a different piece
which puts the king in check.

Now, check and checkmate are defined in the Game class, and
they are passed the is_white value of the team under attack.